- Fix: UB cannot be compiled - Fixed a few compiler warnings
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@8176 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Merc Entering.cpp
@@ -1026,7 +1026,7 @@ void HandleFirstHeliDropOfGame( )
 	CharacterDialogueWithSpecialEvent( 0, 0, 0, DIALOGUE_TACTICAL_UI , FALSE , FALSE , DIALOGUE_SPECIAL_EVENT_ENABLE_AI ,0, 0 );
 }
 
-UINT16 SpawnAirDropElite( INT32 sGridNo )
+UINT8 SpawnAirDropElite( INT32 sGridNo )
 {
 	SOLDIERTYPE *pSoldier;
